+++ splat.yaml
@@ -398,6 +398,7 @@ segments:
     exclusive_ram_id: level
     subsegments:
     - [0x5F050, c, block]
+    - [0x60860, c]
     - [0x60D10, c]
     - [0x617A0, c]
     - [0x617F0, c]
@@ -406,12 +407,13 @@ segments:
     - [0x62010, c]
     - [0x64270, c]
     - [0x64280, .data, block]
-    - [0x642C0, data]
+    - [0x642C0, .data, 60860]
+    - [0x642D0, .data, 60D10]
+    - [0x642E0, .data, 617F0]
     - [0x642F0, .data, 62010]
     - [0x64340, .rodata, block]
     - [0x64380, .rodata, 60D10]
     - [0x64390, .rodata, 617F0]
-    - [0x643A0, rodata]
     - [0x643A0, .rodata, 62010]
     - [0x64420, rodata]
     - { start: auto, name: block, type: .bss, vram: 0x800F5A00 }
